add record/play audio

+++ ScaryBugs/RWTRateView.m
@@ -131,7 +131,7 @@
     if (!self.editable) return;
     
     int newRating = 0;
-    for(int i = self.imageViews.count - 1; i >= 0; i--) {
+    for(int i = (int)self.imageViews.count - 1; i >= 0; i--) {
         UIImageView *imageView = [self.imageViews objectAtIndex:i];        
         if (touchLocation.x > imageView.frame.origin.x) {
             newRating = i+1;
